Auth and start buss

+++ Wazefa.Services/UserServices/UserService.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using Azure;
+using Microsoft.AspNetCore.Identity;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -16,45 +17,62 @@ namespace Wazefa.Services.UserServices
     public class UserService : IUserService
     {
         private readonly IUnitOfWork _unitOfWork;
+        private readonly UserManager<User> _userManager;
+        private readonly RoleManager<IdentityRole> _roleManager;
         private readonly IMapper _mapper;
-        public UserService(IUnitOfWork unitOfWork, IMapper mapper)
+        public UserService(IUnitOfWork unitOfWork, IMapper mapper, UserManager<User> userManager, RoleManager<IdentityRole> roleManager)
         {
             _unitOfWork = unitOfWork;
             _mapper = mapper;
+            _userManager = userManager;
+            _roleManager = roleManager;
         }
-        public async Task<UserResponse?> AddAsync(AddUserRequest dto)
+        public async Task<ResponseResultDto<UserResponse>> AddAsync(AddUserRequest dto)
         {
-            User? addedUser = await _unitOfWork.userRepository.AddAsync(_mapper.Map<User>(dto));
-            await _unitOfWork.SaveAsync();
-            return _mapper.Map<UserResponse>(addedUser);
+            var response = new ResponseResultDto<UserResponse>();
+            User userToAdd = _mapper.Map<User>(dto);
+            var result = await _userManager.CreateAsync(userToAdd, dto.Password);
+            if(await _roleManager.FindByIdAsync(dto.RoleName) == null)
+                await _roleManager.CreateAsync(new IdentityRole { Name = dto.RoleName});
+            await _userManager.AddToRoleAsync(userToAdd, dto.RoleName);
+            return response.MappingResponse(_mapper.Map<UserResponse>(userToAdd));
         }
-        public async Task<UserResponse?> GetByIdAsync(string id)
+        //public async Task<ResponseResultDto<UserResponse>> GetPagedAsync(string id)
+        //{
+        //    var response = new ResponseResultDto<UserResponse>();
+        //    IEnumerable<User> userList = await _unitOfWork.userRepository.GetList();
+        //    if (user == null)
+        //        return response.MappingResponse();
+        //    return response.MappingResponse(_mapper.Map<UserResponse>(user));
+        //}
+        public async Task<ResponseResultDto<UserResponse>> GetByIdAsync(string id)
         {
+            var response = new ResponseResultDto<UserResponse>();
             User? user = await _unitOfWork.userRepository.GetByIdAsync(id);
             if (user == null)
-                return null;
-
-            return _mapper.Map<UserResponse>(user);
+                return response.MappingResponse();
+            return response.MappingResponse(_mapper.Map<UserResponse>(user));
         }
-        public async Task<UserResponse?> UpdateAsync(UpdateUserRequest dto)
+        public async Task<ResponseResultDto<UserResponse>> UpdateAsync(UpdateUserRequest dto)
         {
-
-            User? user = await _unitOfWork.userRepository.GetByIdAsync(dto.Id);
+            var response = new ResponseResultDto<UserResponse>();
+            User? user = await _userManager.FindByIdAsync(dto.Id);
             if (user == null)
-                return null;
-            user = _mapper.Map<User>(dto);
-            User updatedUser = _unitOfWork.userRepository.Update(user);
-            await _unitOfWork.SaveAsync();
-            return _mapper.Map<UserResponse>(user);
+                return response.MappingResponse();
+            _mapper.Map(dto, user);
+            user.ModificationDate = DateTime.Now;
+            IdentityResult updatedUser = await _userManager.UpdateAsync(user);
+            return response.MappingResponse(_mapper.Map<UserResponse>(user));
         }
-        public async Task<bool> DeleteAsync(string id)
+        public async Task<ResponseResultDto<bool>> DeleteAsync(string id)
         {
+            var response = new ResponseResultDto<bool>();
             User? user = await _unitOfWork.userRepository.GetByIdAsync(id);
-            //if (user == null)
-            //    return null;
+            if (user == null)
+                return response.MappingResponse();
             _unitOfWork.userRepository.Delete(user);
             bool isDeleted = await _unitOfWork.SaveAsync() > 0 ? true : false;
-            return isDeleted;
+            return response.MappingResponse(isDeleted);
         }
     }
 }
